studio-backend/requirements.txt
# Core Django framework
Django==5.1.2
djangorestframework==3.15.2

# CORS handling for frontend communication
django-cors-headers==4.5.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Google OAuth
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# HTTP client for Ollama API calls
httpx==0.27.0

# Official Ollama Python client
ollama==0.3.3

# Development dependencies
python-decouple==3.8  # For environment variable management

# Optional: PostgreSQL support (for production)
psycopg2-binary==2.9.9

# Optional: Redis support (for caching/sessions)
redis==5.0.0
django-redis==5.4.0
